Moved RuleAction to separate module
